WaveGenerator - Potentially Finished
Wrote a function that generates a sinewave for a given length (e.g. for
given 'x' it will give y values at array[y]).

Then wrote a few conversion functions and the algorithm which converts
that sinewave data into 2d char[8][8] array, whose bit values represent
the sine-data.

Hopefully this will now be direct "plug-and-play" into the LCD module.

Tested on Processing - seems to work ok - still needs to be tested with
the LCD.